% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BioMedR-datalist.R
\docType{data}
\name{apfp}
\alias{apfp}
\title{Frequent Atom Pairs}
\usage{
data(apfp)
}
\description{
Frequent Atom Pairs
}
\details{
Object stores 4096 most frequent atom pairs generated from DrugBank compounds.
}
\examples{
data(apfp)
}
\keyword{apfp}

