% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BioMedR-datalist.R
\docType{data}
\name{atomprop}
\alias{atomprop}
\title{Standard atomic weights}
\usage{
data(atomprop)
}
\description{
Standard atomic weights
}
\details{
Data frame with atom names, symbols, standard atomic weights,
group number and period number.
}
\examples{
data(atomprop)
}
\keyword{atomprop}

