% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioMedR-datalist.R
\docType{data}
\name{bcl}
\alias{bcl}
\title{2D descriptors of bcl2}
\usage{
data(bcl)
}
\description{
2D descriptors of bcl2
}
\details{
object stores 380 molecules and 160 2D descriptors
}
\examples{
data(bcl)

}
\keyword{bcl}
\keyword{datasets}
