% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001-BioMedR-package.R
\docType{package}
\name{BioMedR-package}
\alias{BioMedR-package}
\alias{BioMedR}
\title{Toolkit for Compound-Protein Interaction in Drug Discovery}
\description{
The BioMedR package offers an R/Bioconductor package emphasizing 
the comprehensive integration of bioinformatics and chemoinformatics 
into a molecular informatics platform for drug discovery.
}
\details{
The comprehensive user guide could be opened with \code{vignette('BioMedR')},
which explains the functionalities included in this package in detail.
The BioMedR package is developed by Computational Biology and 
Drug Design (CBDD) Group, Central South University 
(\url{http://www.scbdd.com/}).

\tabular{ll}{ Package: \tab BioMedR\cr Type: \tab Package\cr
Version: \tab Release 3\cr License: \tab Artistic-2.0\cr }
}
\note{
Bug reports and feature requests should be sent to
\url{https://github.com/wind22zhu/BioMedR/issues}.
}
\examples{
NULL
}
\author{
Minfeng Zhu <\email{wind2zhu@163.com.com}>
        Dongsheng Cao <\email{oriental-cds@163.com}>
}
\keyword{BioMedR}
