% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/103-getDNAGenbank.R
\name{BMgetDNAGenBank}
\alias{BMgetDNAGenBank}
\title{Get DNA/RNA Sequences from Genbank by GI ID}
\usage{
BMgetDNAGenBank(id)
}
\arguments{
\item{id}{A character vector, as the GI ID(s).}
}
\value{
A list, each component contains one of the DNA/RNA sequences.
}
\description{
Get DNA/RNA Sequences from Genbank by GI ID
}
\details{
This function get DNA/RNA sequences from Genbank by GI ID(s).
}
\examples{
\donttest{
# Network latency may slow down this example
# Only test this when your connection is fast enough

ids = c(2, 11)
BMgetDNAGenBank(ids)}
}
\seealso{
See \code{\link{readFASTA}} for reading FASTA format files.
}
\author{
Min-feng Zhu <\email{wind2zhu@163.com}>
}
\keyword{Genbank}
