% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1004-clusterStat.R
\name{clusterStat}
\alias{clusterStat}
\title{generate statistics on sizes of clusters}
\usage{
clusterStat(cls, cluster.result = 1)
}
\arguments{
\item{cls}{The clustering result returned by 'clusterCMP'.}

\item{cluster.result}{If multiple cutoff values are used in clustering process,
                       this argument tells which cutoff value is to be considered here.}
}
\value{
Returns A data frame of two columns.
}
\description{
'cluster.sizestat' is used to do simple statistics on sizes of clusters
generated by 'clusterCMP'. It will return a dataframe which maps a cluster
size to  the number of clusters with that size. It is often used along
with 'cluster.visualize'.
}
\details{
'cluster.sizestat' depends on the format that is returned by 'clusterCMP' - it
will treat the first column as the indecies, and the second column as the
cluster sizes of effective clustering. Because of this, when multiple
cutoffs are used when 'clusterCMP' is called, 'cluster.sizestat' will only
consider the clustering result of the first cutoff. If you want to work on
an alternative cutoff, you have to manually reorder/remove columns.
}
\examples{
data(sdfbcl)
apbcl <- convSDFtoAP(sdfbcl)

cluster <- clusterCMP(db = apbcl, cutoff = c(0.65, 0.5))
clusterStat(cluster[, c(1, 2, 3)])
clusterStat(cluster[, c(1, 4, 5)])
}
\author{
Min-feng Zhu <\email{wind2zhu@163.com}>
}
\references{
...
}
\keyword{clusterStat}

