% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/607-extractDNATCC.R
\name{extrDNATCC}
\alias{extrDNATCC}
\title{The Trinucleotide-based Cross Covariance Descriptor}
\usage{
extrDNATCC(x, index = c("Dnase I", "Nucleosome"), nlag = 2,
  customprops = NULL, normaliztion = FALSE)
}
\arguments{
\item{x}{the input data, which should be a list or file type.}

\item{index}{the physicochemical indices, it should be a list and there are 12
             different physicochemical indices (Table 2), which the users can choose.}

\item{nlag}{an integer larger than or equal to 0 and less than or equal to L-2 (L means the length
            of the shortest DNA sequence in the dataset). It represents the distance between two dinucleotides.}

\item{customprops}{the users can use their own indices to generate the feature vector. It should be a dict,
                   the key is dinucleotide (string), and its corresponding value is a list type.}

\item{normaliztion}{with this option, the final feature vector will be normalized based
                 on the total occurrences of all kmers. Therefore, the elements in the feature vectors
                 represent the frequencies of kmers. The default value of this parameter is False.}
}
\value{
A vector
}
\description{
The Trinucleotide-based Cross Covariance Descriptor
}
\details{
This function calculates the trinucleotide-based cross covariance Descriptor
}
\note{
if the user defined physicochemical indices have not been normalized, it should be normalized.
}
\examples{
x = 'GACTGAACTGCACTTTGGTTTCATATTATTTGCTC'
extrDNATCC(x)
}
\author{
Min-feng Zhu <\email{wind2zhu@163.com}>
}
\seealso{
See \code{\link{extrDNATAC}} and \code{\link{extrDNATACC}}
}
\keyword{TCC}
\keyword{extract}

