% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/427-extractDrugHybridizationRatio.R
\name{extrDrugHybridizationRatio}
\alias{extrDrugHybridizationRatio}
\title{Descriptor that Characterizing Molecular Complexity
in Terms of Carbon Hybridization States}
\usage{
extrDrugHybridizationRatio(molecules, silent = TRUE)
}
\arguments{
\item{molecules}{Parsed molucule object.}

\item{silent}{Logical. Whether the calculating process
should be shown or not, default is \code{TRUE}.}
}
\value{
A data frame, each row represents one of the molecules,
each column represents one feature.
This function returns one column named \code{HybRatio}.
}
\description{
Descriptor that Characterizing Molecular Complexity
in Terms of Carbon Hybridization States
}
\details{
This descriptor calculates the fraction of sp3 carbons to sp2 carbons.
Note that it only considers carbon atoms and rather than use a simple
ratio it reports the value of Nsp3/(Nsp3 + Nsp2).
The original form of the descriptor (i.e., simple ratio)
has been used to characterize molecular complexity,
especially in the are of natural products,
which usually have a high value of the sp3 to sp2 ratio.
}
\examples{
smi = system.file('vignettedata/test.smi', package = 'BioMedR')
mol = readMolFromSmi(smi, type = 'mol')
dat = extrDrugHybridizationRatio(mol)
head(dat)
}
\author{
Min-feng Zhu <\email{wind2zhu@163.com}>,
        Nan Xiao <\url{http://r2s.name}>
}
\keyword{Hybridization}
\keyword{Ratio}
\keyword{extrDrugHybridizationRatio}

