% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/478-extractDrugKR.R
\name{extrDrugKR}
\alias{extrDrugKR}
\title{Calculate the KR (Klekota and Roth) Molecular Fingerprints (in Compact Format)}
\usage{
extrDrugKR(molecules, silent = TRUE)
}
\arguments{
\item{molecules}{Parsed molucule object.}

\item{silent}{Logical. Whether the calculating process 
should be shown or not, default is \code{TRUE}.}
}
\value{
A list, each component represents one of the molecules, each element 
in the component represents the index of which element in the fingerprint is 1.
Each component's name is the length of the fingerprints.
}
\description{
Calculate the KR (Klekota and Roth) Molecular Fingerprints (in Compact Format)
}
\details{
Calculate the 4860 bit fingerprint defined by Klekota and Roth.
}
\examples{
\donttest{
smi = system.file('vignettedata/test.smi', package = 'BioMedR')
mol = readMolFromSmi(smi, type = 'mol')
fp  = extrDrugKR(mol)
head(fp)}

}
\seealso{
\link{extrDrugKRComplete}
}
\author{
Min-feng Zhu <\email{wind2zhu@163.com}>, 
        Nan Xiao <\url{http://r2s.name}>
}
\keyword{extrDrugKR}
