% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/481-extractDrugOBFP3.R
\name{extrDrugOBFP3}
\alias{extrDrugOBFP3}
\title{Calculate the FP3 Molecular Fingerprints}
\usage{
extrDrugOBFP3(molecules, type = c("smile", "sdf"))
}
\arguments{
\item{molecules}{R character string object containing the molecules.
See the example section for details.}

\item{type}{\code{'smile'} or \code{'sdf'}.}
}
\value{
A matrix. Each row represents one molecule,
the columns represent the fingerprints.
}
\description{
Calculate the FP3 Molecular Fingerprints
}
\details{
Calculate the 64 bit FP3 fingerprints provided by OpenBabel.
}
\examples{
\donttest{
mol1 = 'C1CCC1CC(CN(C)(C))CC(=O)CC'  # one molecule SMILE in a vector
mol2 = c('CCC', 'CCN', 'CCN(C)(C)', 'c1ccccc1Cc1ccccc1',
         'C1CCC1CC(CN(C)(C))CC(=O)CC')  # multiple SMILEs in a vector

smifp0 = extrDrugOBFP3(mol1, type = 'smile')
smifp1 = extrDrugOBFP3(mol2, type = 'smile')}
}
\author{
Min-feng Zhu <\email{wind2zhu@163.com}>,
        Nan Xiao <\url{http://r2s.name}>
}
\keyword{extrDrugOBFP3}

