% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/482-extractDrugOBFP4.R
\name{extrDrugOBFP4}
\alias{extrDrugOBFP4}
\title{Calculate the FP4 Molecular Fingerprints}
\usage{
extrDrugOBFP4(molecules, type = c("smile", "sdf"))
}
\arguments{
\item{molecules}{R character string object containing the molecules.
See the example section for details.}

\item{type}{\code{'smile'} or \code{'sdf'}.}
}
\value{
A matrix. Each row represents one molecule,
the columns represent the fingerprints.
}
\description{
Calculate the FP4 Molecular Fingerprints
}
\details{
Calculate the 512 bit FP4 fingerprints provided by OpenBabel.
}
\examples{
mol1 = 'C1CCC1CC(CN(C)(C))CC(=O)CC'  # one molecule SMILE in a vector
mol2 = c('CCC', 'CCN', 'CCN(C)(C)', 'c1ccccc1Cc1ccccc1',
         'C1CCC1CC(CN(C)(C))CC(=O)CC')  # multiple SMILEs in a vector

smifp0 = extrDrugOBFP4(mol1, type = 'smile')
smifp1 = extrDrugOBFP4(mol2, type = 'smile')
}
\author{
Min-feng Zhu <\email{wind2zhu@163.com}>,
        Nan Xiao <\url{http://r2s.name}>
}
\keyword{extrDrugOBFP4}

