% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/510-extractProtCTriad.R
\name{extrProtCTriad}
\alias{extrProtCTriad}
\title{Conjoint Triad Descriptor}
\usage{
extrProtCTriad(x)
}
\arguments{
\item{x}{A character vector, as the input protein sequence.}
}
\value{
A length 343 named vector
}
\description{
Conjoint Triad Descriptor
}
\details{
This function calculates the Conjoint Triad descriptor (Dim: 343).
}
\examples{
x = readFASTA(system.file('protseq/P00750.fasta', package = 'BioMedR'))[[1]]
extrProtCTriad(x)

}
\references{
J.W. Shen, J. Zhang, X.M. Luo, W.L. Zhu, 
K.Q. Yu, K.X. Chen, Y.X. Li, H.L. Jiang. 
Predicting Protein-protein Interactions Based Only on Sequences Information. 
\emph{Proceedings of the National Academy of Sciences}. 007, 104, 4337--4341.
}
\author{
Min-feng Zhu <\email{wind2zhu@163.com}>, 
        Nan Xiao <\url{http://r2s.name}>
}
\keyword{CTriad}
\keyword{Conjoint}
\keyword{Triad}
\keyword{extrProtCTriad}
\keyword{extract}
