% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/511-extractProtSOCN.R
\name{extrProtSOCN}
\alias{extrProtSOCN}
\title{Sequence-Order-Coupling Numbers}
\usage{
extrProtSOCN(x, nlag = 30)
}
\arguments{
\item{x}{A character vector, as the input protein sequence.}

\item{nlag}{The maximum lag, defualt is 30.}
}
\value{
A length \code{nlag * 2} named vector
}
\description{
Sequence-Order-Coupling Numbers
}
\details{
This function calculates the Sequence-Order-Coupling Numbers
(Dim: \code{nlag * 2}, default is 60).
}
\examples{
x = readFASTA(system.file('protseq/P00750.fasta', package = 'BioMedR'))[[1]]
extrProtSOCN(x)
}
\author{
Min-feng Zhu <\email{wind2zhu@163.com}>,
        Nan Xiao <\url{http://r2s.name}>
}
\references{
Kuo-Chen Chou. Prediction of Protein Subcellar Locations by
Incorporating Quasi-Sequence-Order Effect.
\emph{Biochemical and Biophysical Research Communications},
2000, 278, 477-483.

Kuo-Chen Chou and Yu-Dong Cai. Prediction of Protein Sucellular Locations by
GO-FunD-PseAA Predictor.
\emph{Biochemical and Biophysical Research Communications},
2004, 320, 1236-1239.

Gisbert Schneider and Paul Wrede. The Rational Design of
Amino Acid Sequences by Artifical Neural Networks and Simulated
Molecular Evolution: Do Novo Design of an Idealized Leader Cleavge Site.
\emph{Biophys Journal}, 1994, 66, 335-344.
}
\seealso{
See \code{\link{extrProtQSO}} for
quasi-sequence-order descriptors.
}
\keyword{Coupling}
\keyword{Number}
\keyword{Order}
\keyword{SOCN}
\keyword{Sequence}
\keyword{extrProtSOCN}
\keyword{extract}

