% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/702-readPDB.R
\name{readPDB}
\alias{readPDB}
\title{Read Protein Sequences in PDB Format}
\usage{
readPDB(file = system.file("protseq/4HHB.pdb", package = "BioMedR"))
}
\arguments{
\item{file}{The name of the file which the sequences in PDB format are
              to be read from. If it does not contain an absolute or
              relative path, the file name is relative to the current
              working directory, \code{\link{getwd}}.
              The default here is to read the \code{4HHB.PDB} file which
              is present in the \code{protseq} directory of the BioMedR package.}
}
\value{
A character vector, representing the amino acid sequence
        of the single-letter code.
}
\description{
Read Protein Sequences in PDB Format
}
\details{
This function reads protein sequences in PDB (Protein Data Bank) format,
and return the amino acid sequences represented by single-letter code.
}
\examples{
Seq1atp = readPDB(system.file('protseq/1atp.pdb', package = 'BioMedR'))
Seq1atp
}
\author{
Min-feng Zhu <\email{wind2zhu@163.com}>,
        Nan Xiao <\url{http://r2s.name}>
}
\references{
Protein Data Bank Contents Guide: Atomic Coordinate Entry Format Description,
Version 3.30. Accessed 2013-06-26.
\url{ftp://ftp.wwpdb.org/pub/pdb/doc/format_descriptions/Format_v33_Letter.pdf}
}
\seealso{
See \code{\link{readFASTA}} for reading protein sequences
         in FASTA format.
}
\keyword{BioMedR}
\keyword{PDB}
\keyword{readPDB}

