% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/708-revchars.R
\name{revchars}
\alias{revchars}
\title{The Reverse chars}
\usage{
revchars(x)
}
\arguments{
\item{x}{the input data, which should be a string.}
}
\value{
A vector
}
\description{
The Reverse chars
}
\details{
This function calculates Reverse chars
}
\note{
if the user defined physicochemical indices have not been normalized, it should be normalized.
}
\examples{
 
x = 'GACTGAACTGCACTTTGGTTTCATATTATTTGCTC'
revchars(x)

}
\author{
Min-feng Zhu <\email{wind2zhu@163.com}>
}
\keyword{extract}
\keyword{reverse_chars}
