% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioMedR-datalist.R
\docType{data}
\name{sdfbcl}
\alias{sdfbcl}
\title{SD file in \code{SDFset} object}
\usage{
data(sdfbcl)
}
\description{
SD file in \code{SDFset} object
}
\details{
\code{SDFset} object for the 50 SDF.
}
\examples{
data(sdfbcl)

}
\keyword{datasets}
\keyword{sdf}
\keyword{sdfbcl}
