% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioMedR-datalist.R
\docType{data}
\name{som.bcl}
\alias{som.bcl}
\title{kohonen object}
\usage{
data(som.bcl)
}
\description{
kohonen object.
}
\details{
kohonen object
}
\examples{
data(som.bcl)

}
\keyword{som.bcl}
\keyword{datasets}
