% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/600-make_kmer_index.R
\name{make_kmer_index}
\alias{make_kmer_index}
\title{Calculate The Basic Kmer Feature Vector}
\usage{
make_kmer_index(k, alphabet = "ACGT")
}
\arguments{
\item{k}{the k value of kmer, it should be an integer larger than 0.}

\item{alphabet}{the}
}
\value{
The result character vector
}
\description{
Calculate The Basic Kmer Feature Vector
}
\details{
This function calculate the basic kmer feature vector.
}
\examples{
make_kmer_index(2, alphabet = "ACGT")
}
\seealso{
See \code{\link{extrDNAkmer}}
}
\author{
Min-feng Zhu <\email{wind2zhu@163.com}>
}
\keyword{index}
\keyword{kmer}
